Reorganize sidepanel components, select journey from timetable WIP.
